Added marks on the map for earthquakes coming from the default API call.
