+++ app/src/main/java/com/codepath/earthquakemonitor/MapActivity.java
@@ -36,7 +36,11 @@ import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.SupportMapFragment;
+import com.google.android.gms.maps.model.BitmapDescriptor;
+import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.Marker;
+import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 
@@ -47,19 +51,19 @@ import static com.google.android.gms.location.LocationServices.getFusedLocationP
 
 import android.support.v7.widget.Toolbar;
 
-@RuntimePermissions
+//@RuntimePermissions
 public class MapActivity extends AppCompatActivity
     implements EarthquakeListFragment.EarthquakeSelectedListener
 {
 
-    private SupportMapFragment mapFragment;
-    private GoogleMap map;
-    private LocationRequest mLocationRequest;
-    Location mCurrentLocation;
-    private long UPDATE_INTERVAL = 60000;  /* 60 secs */
-    private long FASTEST_INTERVAL = 5000; /* 5 secs */
-
-    private final static String KEY_LOCATION = "location";
+//    private SupportMapFragment mapFragment;
+//    private GoogleMap map;
+//    private LocationRequest mLocationRequest;
+//    Location mCurrentLocation;
+//    private long UPDATE_INTERVAL = 60000;  /* 60 secs */
+//    private long FASTEST_INTERVAL = 5000; /* 5 secs */
+//
+//    private final static String KEY_LOCATION = "location";
 
     /*
      * Define a request code to send to Google Play services This code is
@@ -78,28 +82,15 @@ public class MapActivity extends AppCompatActivity
         // Make sure the toolbar exists in the activity and is not null
         setSupportActionBar(toolbar);
 
-        if (TextUtils.isEmpty(getResources().getString(R.string.google_maps_api_key))) {
-            throw new IllegalStateException("You forgot to supply a Google Maps API key");
-        }
+//        if (TextUtils.isEmpty(getResources().getString(R.string.google_maps_api_key))) {
+//            throw new IllegalStateException("You forgot to supply a Google Maps API key");
+//        }
 
-        if (savedInstanceState != null && savedInstanceState.keySet().contains(KEY_LOCATION)) {
-            // Since KEY_LOCATION was found in the Bundle, we can be sure that mCurrentLocation
-            // is not null.
-            mCurrentLocation = savedInstanceState.getParcelable(KEY_LOCATION);
-        }
-
-        // Create the map fragment
-        mapFragment = ((SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map));
-        if (mapFragment != null) {
-            mapFragment.getMapAsync(new OnMapReadyCallback() {
-                @Override
-                public void onMapReady(GoogleMap map) {
-                    loadMap(map);
-                }
-            });
-        } else {
-            Toast.makeText(this, "Error - Map Fragment was null!!", Toast.LENGTH_SHORT).show();
-        }
+//        if (savedInstanceState != null && savedInstanceState.keySet().contains(KEY_LOCATION)) {
+//            // Since KEY_LOCATION was found in the Bundle, we can be sure that mCurrentLocation
+//            // is not null.
+//            mCurrentLocation = savedInstanceState.getParcelable(KEY_LOCATION);
+//        }
 
         //TODO Create the earthquake list fragment
         // Create the user fragment
@@ -144,177 +135,188 @@ public class MapActivity extends AppCompatActivity
         return super.onOptionsItemSelected(item);
     }
 
-    protected void loadMap(GoogleMap googleMap) {
-        map = googleMap;
-        if (map != null) {
-            // Map is ready
-            Toast.makeText(this, "Map Fragment was loaded properly!", Toast.LENGTH_SHORT).show();
-            MapActivityPermissionsDispatcher.getMyLocationWithCheck(this);
-            MapActivityPermissionsDispatcher.startLocationUpdatesWithCheck(this);
-        } else {
-            Toast.makeText(this, "Error - Map was null!!", Toast.LENGTH_SHORT).show();
-        }
-    }
+//    protected void loadMap(GoogleMap googleMap) {
+//        map = googleMap;
+//        if (map != null) {
+//            // Map is ready
+//            Toast.makeText(this, "Map Fragment was loaded properly!", Toast.LENGTH_SHORT).show();
+//            MapActivityPermissionsDispatcher.getMyLocationWithCheck(this);
+//            MapActivityPermissionsDispatcher.startLocationUpdatesWithCheck(this);
+//        } else {
+//            Toast.makeText(this, "Error - Map was null!!", Toast.LENGTH_SHORT).show();
+//        }
+//    }
 
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        MapActivityPermissionsDispatcher.onRequestPermissionsResult(this, requestCode, grantResults);
-    }
-
-    @NeedsPermission({Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION})
-    void getMyLocation() {
-        //noinspection MissingPermission
-        map.setMyLocationEnabled(true);
-
-        FusedLocationProviderClient locationClient = getFusedLocationProviderClient(this);
-        //noinspection MissingPermission
-        locationClient.getLastLocation()
-                .addOnSuccessListener(new OnSuccessListener<Location>() {
-                    @Override
-                    public void onSuccess(Location location) {
-                        if (location != null) {
-                            onLocationChanged(location);
-                        }
-                    }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        Log.d("MapActivity", "Error trying to get last GPS location");
-                        e.printStackTrace();
-                    }
-                });
-    }
-
-    /*
-     * Called when the Activity becomes visible.
-    */
-    @Override
-    protected void onStart() {
-        super.onStart();
-    }
-
-    /*
-     * Called when the Activity is no longer visible.
-	 */
-    @Override
-    protected void onStop() {
-        super.onStop();
-    }
-
-    private boolean isGooglePlayServicesAvailable() {
-        // Check that Google Play services is available
-        int resultCode = GooglePlayServicesUtil.isGooglePlayServicesAvailable(this);
-        // If Google Play services is available
-        if (ConnectionResult.SUCCESS == resultCode) {
-            // In debug mode, log the status
-            Log.d("Location Updates", "Google Play services is available.");
-            return true;
-        } else {
-            // Get the error dialog from Google Play services
-            Dialog errorDialog = GooglePlayServicesUtil.getErrorDialog(resultCode, this,
-                    CONNECTION_FAILURE_RESOLUTION_REQUEST);
-
-            // If Google Play services can provide an error dialog
-            if (errorDialog != null) {
-                // Create a new DialogFragment for the error dialog
-                ErrorDialogFragment errorFragment = new ErrorDialogFragment();
-                errorFragment.setDialog(errorDialog);
-                errorFragment.show(getSupportFragmentManager(), "Location Updates");
-            }
-
-            return false;
-        }
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-
-        // Display the connection status
-
-        if (mCurrentLocation != null) {
-            Toast.makeText(this, "GPS location was found!", Toast.LENGTH_SHORT).show();
-            LatLng latLng = new LatLng(mCurrentLocation.getLatitude(), mCurrentLocation.getLongitude());
-            CameraUpdate cameraUpdate = CameraUpdateFactory.newLatLngZoom(latLng, 17);
-            map.animateCamera(cameraUpdate);
-        } else {
-            Toast.makeText(this, "Current location was null, enable GPS on emulator!", Toast.LENGTH_SHORT).show();
-        }
-        MapActivityPermissionsDispatcher.startLocationUpdatesWithCheck(this);
-    }
-
-    @NeedsPermission({Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION})
-    protected void startLocationUpdates() {
-        mLocationRequest = new LocationRequest();
-        mLocationRequest.setPriority(LocationRequest.PRIORITY_BALANCED_POWER_ACCURACY);
-        mLocationRequest.setInterval(UPDATE_INTERVAL);
-        mLocationRequest.setFastestInterval(FASTEST_INTERVAL);
-
-        LocationSettingsRequest.Builder builder = new LocationSettingsRequest.Builder();
-        builder.addLocationRequest(mLocationRequest);
-        LocationSettingsRequest locationSettingsRequest = builder.build();
-
-        SettingsClient settingsClient = LocationServices.getSettingsClient(this);
-        settingsClient.checkLocationSettings(locationSettingsRequest);
-        //noinspection MissingPermission
-        getFusedLocationProviderClient(this).requestLocationUpdates(mLocationRequest, new LocationCallback() {
-                    @Override
-                    public void onLocationResult(LocationResult locationResult) {
-                        onLocationChanged(locationResult.getLastLocation());
-                    }
-                },
-                Looper.myLooper());
-    }
-
-    public void onLocationChanged(Location location) {
-        // GPS may be turned off
-        if (location == null) {
-            return;
-        }
-
-        // Report to the UI that the location was updated
-
-        mCurrentLocation = location;
-        String msg = "Updated Location: " +
-                Double.toString(location.getLatitude()) + "," +
-                Double.toString(location.getLongitude());
-        Toast.makeText(this, msg, Toast.LENGTH_SHORT).show();
-    }
-
-    public void onSaveInstanceState(Bundle savedInstanceState) {
-        savedInstanceState.putParcelable(KEY_LOCATION, mCurrentLocation);
-        super.onSaveInstanceState(savedInstanceState);
-    }
-
-    // Define a DialogFragment that displays the error dialog
-    public static class ErrorDialogFragment extends android.support.v4.app.DialogFragment {
-
-        // Global field to contain the error dialog
-        private Dialog mDialog;
-
-        // Default constructor. Sets the dialog field to null
-        public ErrorDialogFragment() {
-            super();
-            mDialog = null;
-        }
-
-        // Set the dialog to display
-        public void setDialog(Dialog dialog) {
-            mDialog = dialog;
-        }
-
-        // Return a Dialog to the DialogFragment.
-        @Override
-        public Dialog onCreateDialog(Bundle savedInstanceState) {
-            return mDialog;
-        }
-    }
+//    @Override
+//    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+//        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+//        MapActivityPermissionsDispatcher.onRequestPermissionsResult(this, requestCode, grantResults);
+//    }
+//
+//    @NeedsPermission({Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION})
+//    void getMyLocation() {
+//        //noinspection MissingPermission
+//        map.setMyLocationEnabled(true);
+//
+//        FusedLocationProviderClient locationClient = getFusedLocationProviderClient(this);
+//        //noinspection MissingPermission
+//        locationClient.getLastLocation()
+//                .addOnSuccessListener(new OnSuccessListener<Location>() {
+//                    @Override
+//                    public void onSuccess(Location location) {
+//                        if (location != null) {
+//                            onLocationChanged(location);
+//                        }
+//                    }
+//                })
+//                .addOnFailureListener(new OnFailureListener() {
+//                    @Override
+//                    public void onFailure(@NonNull Exception e) {
+//                        Log.d("MapActivity", "Error trying to get last GPS location");
+//                        e.printStackTrace();
+//                    }
+//                });
+//    }
+//
+//    /*
+//     * Called when the Activity becomes visible.
+//    */
+//    @Override
+//    protected void onStart() {
+//        super.onStart();
+//    }
+//
+//    /*
+//     * Called when the Activity is no longer visible.
+//	 */
+//    @Override
+//    protected void onStop() {
+//        super.onStop();
+//    }
+//
+//    private boolean isGooglePlayServicesAvailable() {
+//        // Check that Google Play services is available
+//        int resultCode = GooglePlayServicesUtil.isGooglePlayServicesAvailable(this);
+//        // If Google Play services is available
+//        if (ConnectionResult.SUCCESS == resultCode) {
+//            // In debug mode, log the status
+//            Log.d("Location Updates", "Google Play services is available.");
+//            return true;
+//        } else {
+//            // Get the error dialog from Google Play services
+//            Dialog errorDialog = GooglePlayServicesUtil.getErrorDialog(resultCode, this,
+//                    CONNECTION_FAILURE_RESOLUTION_REQUEST);
+//
+//            // If Google Play services can provide an error dialog
+//            if (errorDialog != null) {
+//                // Create a new DialogFragment for the error dialog
+//                ErrorDialogFragment errorFragment = new ErrorDialogFragment();
+//                errorFragment.setDialog(errorDialog);
+//                errorFragment.show(getSupportFragmentManager(), "Location Updates");
+//            }
+//
+//            return false;
+//        }
+//    }
+//
+//    @Override
+//    protected void onResume() {
+//        super.onResume();
+//
+//        // Display the connection status
+//
+//        if (mCurrentLocation != null) {
+//            Toast.makeText(this, "GPS location was found!", Toast.LENGTH_SHORT).show();
+//            LatLng latLng = new LatLng(mCurrentLocation.getLatitude(), mCurrentLocation.getLongitude());
+//            CameraUpdate cameraUpdate = CameraUpdateFactory.newLatLngZoom(latLng, 17);
+//            map.animateCamera(cameraUpdate);
+//        } else {
+//            Toast.makeText(this, "Current location was null, enable GPS on emulator!", Toast.LENGTH_SHORT).show();
+//        }
+//        MapActivityPermissionsDispatcher.startLocationUpdatesWithCheck(this);
+//    }
+//
+//    @NeedsPermission({Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION})
+//    protected void startLocationUpdates() {
+//        mLocationRequest = new LocationRequest();
+//        mLocationRequest.setPriority(LocationRequest.PRIORITY_BALANCED_POWER_ACCURACY);
+//        mLocationRequest.setInterval(UPDATE_INTERVAL);
+//        mLocationRequest.setFastestInterval(FASTEST_INTERVAL);
+//
+//        LocationSettingsRequest.Builder builder = new LocationSettingsRequest.Builder();
+//        builder.addLocationRequest(mLocationRequest);
+//        LocationSettingsRequest locationSettingsRequest = builder.build();
+//
+//        SettingsClient settingsClient = LocationServices.getSettingsClient(this);
+//        settingsClient.checkLocationSettings(locationSettingsRequest);
+//        //noinspection MissingPermission
+//        getFusedLocationProviderClient(this).requestLocationUpdates(mLocationRequest, new LocationCallback() {
+//                    @Override
+//                    public void onLocationResult(LocationResult locationResult) {
+//                        onLocationChanged(locationResult.getLastLocation());
+//                    }
+//                },
+//                Looper.myLooper());
+//    }
+//
+//    public void onLocationChanged(Location location) {
+//        // GPS may be turned off
+//        if (location == null) {
+//            return;
+//        }
+//
+//        // Report to the UI that the location was updated
+//
+//        mCurrentLocation = location;
+//        String msg = "Updated Location: " +
+//                Double.toString(location.getLatitude()) + "," +
+//                Double.toString(location.getLongitude());
+//        Toast.makeText(this, msg, Toast.LENGTH_SHORT).show();
+//    }
+//
+//    public void onSaveInstanceState(Bundle savedInstanceState) {
+//        savedInstanceState.putParcelable(KEY_LOCATION, mCurrentLocation);
+//        super.onSaveInstanceState(savedInstanceState);
+//    }
+//
+//    // Define a DialogFragment that displays the error dialog
+//    public static class ErrorDialogFragment extends android.support.v4.app.DialogFragment {
+//
+//        // Global field to contain the error dialog
+//        private Dialog mDialog;
+//
+//        // Default constructor. Sets the dialog field to null
+//        public ErrorDialogFragment() {
+//            super();
+//            mDialog = null;
+//        }
+//
+//        // Set the dialog to display
+//        public void setDialog(Dialog dialog) {
+//            mDialog = dialog;
+//        }
+//
+//        // Return a Dialog to the DialogFragment.
+//        @Override
+//        public Dialog onCreateDialog(Bundle savedInstanceState) {
+//            return mDialog;
+//        }
+//    }
 
     @Override
     public void onEarthquakeClicked(Earthquake earthquake)
     {
+//        // Set the color of the marker to green
+//        BitmapDescriptor defaultMarker =
+//                BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN);
+//        // listingPosition is a LatLng point
+//        LatLng listingPosition = new LatLng(-33.867, 151.206);
+//        // Create the marker on the fragment
+//        Marker mapMarker = map.addMarker(new MarkerOptions()
+//                .position(listingPosition)
+//                .title("Some title here")
+//                .snippet("Some description here")
+//                .icon(defaultMarker));
         Toast.makeText(getApplicationContext(), "Please launch webpage for details!", Toast.LENGTH_LONG).show();
     }
 }
