+++ app/src/main/java/com/codepath/earthquakemonitor/fragments/EarthquakeListFragment.java
@@ -1,6 +1,11 @@
 package com.codepath.earthquakemonitor.fragments;
 
+import android.Manifest;
+import android.app.Dialog;
+import android.location.Location;
 import android.os.Bundle;
+import android.os.Looper;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.LinearLayoutManager;
@@ -9,12 +14,35 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import com.codepath.earthquakemonitor.Adapters.EarthquakeAdapter;
 import com.codepath.earthquakemonitor.EarthquakeClient;
+import com.codepath.earthquakemonitor.MapActivity;
 import com.codepath.earthquakemonitor.R;
 import com.codepath.earthquakemonitor.models.Earthquake;
 import com.codepath.earthquakemonitor.models.Filters;
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.GooglePlayServicesUtil;
+import com.google.android.gms.location.FusedLocationProviderClient;
+import com.google.android.gms.location.LocationCallback;
+import com.google.android.gms.location.LocationRequest;
+import com.google.android.gms.location.LocationResult;
+import com.google.android.gms.location.LocationServices;
+import com.google.android.gms.location.LocationSettingsRequest;
+import com.google.android.gms.location.SettingsClient;
+import com.google.android.gms.maps.CameraUpdate;
+import com.google.android.gms.maps.CameraUpdateFactory;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.OnMapReadyCallback;
+import com.google.android.gms.maps.SupportMapFragment;
+import com.google.android.gms.maps.model.BitmapDescriptor;
+import com.google.android.gms.maps.model.BitmapDescriptorFactory;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.Marker;
+import com.google.android.gms.maps.model.MarkerOptions;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.loopj.android.http.AsyncHttpResponseHandler;
 import com.loopj.android.http.JsonHttpResponseHandler;
 
@@ -26,18 +54,35 @@ import java.util.ArrayList;
 import java.util.List;
 
 import cz.msebera.android.httpclient.Header;
+import permissions.dispatcher.NeedsPermission;
+import permissions.dispatcher.RuntimePermissions;
+
+import static com.google.android.gms.location.LocationServices.getFusedLocationProviderClient;
 
 /**
  * Created by hezhang on 10/14/17.
  */
 
+@RuntimePermissions
 public class EarthquakeListFragment extends Fragment
     implements EarthquakeAdapter.EarthquakeAdapterListener
 {
     private final String TAG = "EarthquakeListFragment";
-//
-//    private EarthquakeClient client;
-//    Filters filter = Filters.getInstance();
+
+    private SupportMapFragment mapFragment;
+    private GoogleMap map;
+    private LocationRequest mLocationRequest;
+    Location mCurrentLocation;
+    private long UPDATE_INTERVAL = 60000;  /* 60 secs */
+    private long FASTEST_INTERVAL = 5000; /* 5 secs */
+
+    private final static String KEY_LOCATION = "location";
+
+    /*
+     * Define a request code to send to Google Play services This code is
+     * returned in Activity.onActivityResult
+     */
+    private final static int CONNECTION_FAILURE_RESOLUTION_REQUEST = 9000;
 
     public interface EarthquakeSelectedListener {
         public void onEarthquakeClicked(Earthquake earthquake);
@@ -51,6 +96,27 @@ public class EarthquakeListFragment extends Fragment
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState)
     {
+        if (savedInstanceState != null && savedInstanceState.keySet().contains(KEY_LOCATION)) {
+            // Since KEY_LOCATION was found in the Bundle, we can be sure that mCurrentLocation
+            // is not null.
+            mCurrentLocation = savedInstanceState.getParcelable(KEY_LOCATION);
+        }
+
+        // Create the map fragment
+        mapFragment = ((SupportMapFragment) getFragmentManager().findFragmentById(R.id.map));
+        if (mapFragment != null) {
+            mapFragment.getMapAsync(new OnMapReadyCallback() {
+                @Override
+                public void onMapReady(GoogleMap map) {
+                    loadMap(map);
+                }
+            });
+        } else {
+            Toast.makeText(getContext(), "Error - Map Fragment was null!!", Toast.LENGTH_SHORT).show();
+        }
+
+
+
         // inflate the layout
         View v = inflater.inflate(R.layout.fragment_earthquake_list, container, false);
         // Find the recycler view
@@ -65,6 +131,8 @@ public class EarthquakeListFragment extends Fragment
         LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getContext());
         rvEarthquakes.setLayoutManager(linearLayoutManager);
         return v;
+
+
     }
 
     public void addItems(JSONObject response) {
@@ -77,6 +145,17 @@ public class EarthquakeListFragment extends Fragment
                 Earthquake earthquake = Earthquake.fromJson(jsonArray.getJSONObject(i));
                 earthquakes.add(earthquake);
                 earthquakeAdapter.notifyItemInserted(earthquakes.size() - 1);
+                // Set the color of the marker to green
+                BitmapDescriptor defaultMarker =
+                        BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN);
+                // listingPosition is a LatLng point
+                LatLng listingPosition = new LatLng(earthquake.getLatitude(), earthquake.getLongitude());
+                // Create the marker on the fragment
+                Marker mapMarker = map.addMarker(new MarkerOptions()
+                        .position(listingPosition)
+                        .title("Some title here")
+                        .snippet("Some description here")
+                        .icon(defaultMarker));
             }
         }
         catch (JSONException e) {
@@ -89,4 +168,173 @@ public class EarthquakeListFragment extends Fragment
         Earthquake earthquake = earthquakes.get(position);
         ((EarthquakeSelectedListener) getActivity()).onEarthquakeClicked(earthquake);
     }
+
+    protected void loadMap(GoogleMap googleMap) {
+        map = googleMap;
+        if (map != null) {
+            // Map is ready
+            Toast.makeText(getContext(), "Map Fragment was loaded properly!", Toast.LENGTH_SHORT).show();
+            EarthquakeListFragmentPermissionsDispatcher.getMyLocationWithCheck(this);
+            EarthquakeListFragmentPermissionsDispatcher.startLocationUpdatesWithCheck(this);
+        } else {
+            Toast.makeText(getContext(), "Error - Map was null!!", Toast.LENGTH_SHORT).show();
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        EarthquakeListFragmentPermissionsDispatcher.onRequestPermissionsResult(this, requestCode, grantResults);
+    }
+
+    @NeedsPermission({Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION})
+    void getMyLocation() {
+        //noinspection MissingPermission
+        map.setMyLocationEnabled(true);
+
+        FusedLocationProviderClient locationClient = getFusedLocationProviderClient(getActivity());
+        //noinspection MissingPermission
+        locationClient.getLastLocation()
+                .addOnSuccessListener(new OnSuccessListener<Location>() {
+                    @Override
+                    public void onSuccess(Location location) {
+                        if (location != null) {
+                            onLocationChanged(location);
+                        }
+                    }
+                })
+                .addOnFailureListener(new OnFailureListener() {
+                    @Override
+                    public void onFailure(@NonNull Exception e) {
+                        Log.d("MapActivity", "Error trying to get last GPS location");
+                        e.printStackTrace();
+                    }
+                });
+    }
+
+    /*
+     * Called when the Activity becomes visible.
+    */
+    @Override
+    public void onStart() {
+        super.onStart();
+    }
+
+    /*
+     * Called when the Activity is no longer visible.
+	 */
+    @Override
+    public void onStop() {
+        super.onStop();
+    }
+
+    private boolean isGooglePlayServicesAvailable() {
+        // Check that Google Play services is available
+        int resultCode = GooglePlayServicesUtil.isGooglePlayServicesAvailable(getActivity());
+        // If Google Play services is available
+        if (ConnectionResult.SUCCESS == resultCode) {
+            // In debug mode, log the status
+            Log.d("Location Updates", "Google Play services is available.");
+            return true;
+        } else {
+            // Get the error dialog from Google Play services
+            Dialog errorDialog = GooglePlayServicesUtil.getErrorDialog(resultCode, getActivity(),
+                    CONNECTION_FAILURE_RESOLUTION_REQUEST);
+
+            // If Google Play services can provide an error dialog
+            if (errorDialog != null) {
+                // Create a new DialogFragment for the error dialog
+                EarthquakeListFragment.ErrorDialogFragment errorFragment = new EarthquakeListFragment.ErrorDialogFragment();
+                errorFragment.setDialog(errorDialog);
+                errorFragment.show(getFragmentManager(), "Location Updates");
+            }
+
+            return false;
+        }
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+
+        // Display the connection status
+
+        if (mCurrentLocation != null) {
+            Toast.makeText(getContext(), "GPS location was found!", Toast.LENGTH_SHORT).show();
+            LatLng latLng = new LatLng(mCurrentLocation.getLatitude(), mCurrentLocation.getLongitude());
+            CameraUpdate cameraUpdate = CameraUpdateFactory.newLatLngZoom(latLng, 17);
+            map.animateCamera(cameraUpdate);
+        } else {
+            Toast.makeText(getContext(), "Current location was null, enable GPS on emulator!", Toast.LENGTH_SHORT).show();
+        }
+        EarthquakeListFragmentPermissionsDispatcher.startLocationUpdatesWithCheck(this);
+    }
+
+    @NeedsPermission({Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION})
+    protected void startLocationUpdates() {
+        mLocationRequest = new LocationRequest();
+        mLocationRequest.setPriority(LocationRequest.PRIORITY_BALANCED_POWER_ACCURACY);
+        mLocationRequest.setInterval(UPDATE_INTERVAL);
+        mLocationRequest.setFastestInterval(FASTEST_INTERVAL);
+
+        LocationSettingsRequest.Builder builder = new LocationSettingsRequest.Builder();
+        builder.addLocationRequest(mLocationRequest);
+        LocationSettingsRequest locationSettingsRequest = builder.build();
+
+        SettingsClient settingsClient = LocationServices.getSettingsClient(getActivity());
+        settingsClient.checkLocationSettings(locationSettingsRequest);
+        //noinspection MissingPermission
+        getFusedLocationProviderClient(getActivity()).requestLocationUpdates(mLocationRequest, new LocationCallback() {
+                    @Override
+                    public void onLocationResult(LocationResult locationResult) {
+                        onLocationChanged(locationResult.getLastLocation());
+                    }
+                },
+                Looper.myLooper());
+    }
+
+    public void onLocationChanged(Location location) {
+        // GPS may be turned off
+        if (location == null) {
+            return;
+        }
+
+        // Report to the UI that the location was updated
+
+        mCurrentLocation = location;
+        String msg = "Updated Location: " +
+                Double.toString(location.getLatitude()) + "," +
+                Double.toString(location.getLongitude());
+        Toast.makeText(getContext(), msg, Toast.LENGTH_SHORT).show();
+    }
+
+    public void onSaveInstanceState(Bundle savedInstanceState) {
+        savedInstanceState.putParcelable(KEY_LOCATION, mCurrentLocation);
+        super.onSaveInstanceState(savedInstanceState);
+    }
+
+    // Define a DialogFragment that displays the error dialog
+    public static class ErrorDialogFragment extends android.support.v4.app.DialogFragment {
+
+        // Global field to contain the error dialog
+        private Dialog mDialog;
+
+        // Default constructor. Sets the dialog field to null
+        public ErrorDialogFragment() {
+            super();
+            mDialog = null;
+        }
+
+        // Set the dialog to display
+        public void setDialog(Dialog dialog) {
+            mDialog = dialog;
+        }
+
+        // Return a Dialog to the DialogFragment.
+        @Override
+        public Dialog onCreateDialog(Bundle savedInstanceState) {
+            return mDialog;
+        }
+    }
+
 }
